pyproject.toml
[project]
name = "atd-music-search"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dreinon",email = "danielreinon@outlook.com"},
]
dependencies = [
    "fastapi>=0.109.0",
    "python-jose>=3.3.0",
    "gunicorn>=21.2.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "typer[all]>=0.9.0",
    "inquirerpy>=0.3.4",
    "prompt-toolkit>=3.0.43",
]

requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "watchfiles>=0.21.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "pre-commit>=3.6.0",
]

[tool.pdm.scripts]
set_precommit_hooks = "pre-commit install --install-hooks"
format = {shell = "black . && isort ."}
lint = {shell = "pylint src"}
api = {cmd = "python -m scripts.run_local_server", env = {ENV = "config"}}
cli = {cmd = "python -m scripts.run_cli", env = {ENV = "config"}}
server = "gunicorn src.atd_music_search_api.api:app -b 0.0.0.0:8080 -k uvicorn.workers.UvicornWorker -w 4 -t 300 --reload"
tunnel = "ngrok http 8080"


[tool.pdm]
distribution = false

[tool.isort]
profile = "black"
